+++ Example/ZcashLightClientSample/ZcashLightClientSample/Sync Blocks/SyncBlocksViewController.swift
@@ -45,6 +45,12 @@ class SyncBlocksViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
+        self.navigationItem.rightBarButtonItem = UIBarButtonItem(
+            barButtonSystemItem: .trash,
+            target: self,
+            action: #selector(wipe(_:))
+        )
+
         statusLabel.text = textFor(state: synchronizer.status)
         progressBar.progress = 0
         let center = NotificationCenter.default
@@ -339,5 +345,69 @@ extension SDKMetrics.BlockMetricReport: CustomDebugStringConvertible {
             batchSize: \(self.batchSize)
             duration: \(self.duration)
         """
+    }
+}
+
+// MARK: Wipe
+
+extension SyncBlocksViewController {
+    @objc func wipe(_ sender: Any?) {
+        let alert = UIAlertController(
+            title: "Wipe the wallet?",
+            message: """
+            You are about to clear existing databases. All synced blocks, stored TXs, etc will be removed form this device only. If the sync is in
+            progress it may take some time. Please be patient.
+            """,
+            preferredStyle: .alert
+        )
+
+        alert.addAction(
+            UIAlertAction(
+                title: "Drop it like it's FIAT",
+                style: UIAlertAction.Style.destructive
+            ) { [weak self] _ in
+                guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { return }
+                appDelegate.wipe() { [weak self] possibleError in
+                    if let error = possibleError {
+                        self?.wipeFailedAlert(error: error)
+                    } else {
+                        self?.wipeSuccessfullAlert()
+                    }
+                }
+            }
+        )
+
+        alert.addAction(UIAlertAction(title: "No please! Have mercy!", style: UIAlertAction.Style.cancel, handler: nil))
+
+        self.present(alert, animated: true, completion: nil)
+    }
+
+    private func wipeSuccessfullAlert() {
+        let alert = UIAlertController(
+            title: "Wipe is done!",
+            message: nil,
+            preferredStyle: .alert
+        )
+
+        alert.addAction(UIAlertAction(title: "Ok", style: UIAlertAction.Style.default, handler: nil))
+        self.present(alert, animated: true, completion: nil)
+    }
+
+    private func wipeFailedAlert(error: Error) {
+        loggerProxy.error("Wipe error: \(error)")
+
+        let alert = UIAlertController(
+            title: "Wipe FAILED!",
+            message: """
+            Something bad happened and wipe failed. This may happen only when some basic IO disk operations failed. The SDK may end up in \
+            inconsistent state. It's suggested to call the wipe again until it succeeds. Sorry.
+
+            \(error)
+            """,
+            preferredStyle: .alert
+        )
+
+        alert.addAction(UIAlertAction(title: "Ok", style: UIAlertAction.Style.default, handler: nil))
+        self.present(alert, animated: true, completion: nil)
     }
 }